Add remote upload feature Rename functions' name in app.client.remote_backup     downloadRequest -> startDownload     downloadBackup -> getDownloadResult Move open download temp file from startDownload to process block of UPLOAD_DOWNLOAD_START Adjust buffer limit from 4096, 8192 to 65535

+++ app/client/remote_backup.py
@@ -11,13 +11,19 @@ import app.com.packet.login
 import app.com.packet.backup 
 
 class BaksysRemoteBackup:
+    UPLOAD_CHUNK_SIZE = 65535
+    
     def __init__(this):
         this.socket             = BaksysClientSocket()
         this.responseLock       = threading.Event()
         this.remoteResponse     = { }
         this.downloadFile       = { }
+        this.uploadFile         = { }
+        this.uploadInterrupted  = False
+        this.uploadThread       = None
         this.lastOperationError = ''
         this.onDownloadProcess  = BaksysEvent()
+        this.onUploadProcess    = BaksysEvent()
         
     def isConnecting(this):
         return this.socket.isConnecting()
@@ -50,14 +56,77 @@ class BaksysRemoteBackup:
         this.socket.send(packet.backup.listRequest())
         return this.waitForResponse()
         
+    def startUpload(this, path):
+        this.uploadFile = {                                           \
+            'path'       : os.path.join(config['backup_path'], path), \
+            'file'       : None,                                      \
+            'size'       : 0,                                         \
+        }
+        this.socket.send(packet.backup.uploadStartRequest(path))
+        result = this.waitForResponse()
+        if not result:
+            this.uploadInterrupt()
+        return result
+            
+    def getUploadResult(this):
+        def upload():
+            try:
+                with open(this.uploadFile['path'], 'rb') as file:
+                    begin  = file.tell()
+                    file.seek(0, 2)
+                    size   = file.tell() - begin
+                    this.uploadFile['size'] = size
+                    offset = 0
+                    file.seek(0, 0)
+                    while size > 0 and this.socket.isConnecting():
+                        if this.uploadInterrupted:
+                            this.socket.send(packet.backup.uploadBreakResponse())
+                            return
+                        sizeToWrite = min(this.UPLOAD_CHUNK_SIZE, size)
+                        data        = file.read(sizeToWrite)
+                        this.socket.send(packet.backup.uploadDataRequest(offset, data))
+                        offset += sizeToWrite
+                        size   -= sizeToWrite
+                        
+                        # invoke event
+                        args = BaksysEventArgs()
+                        args.progress = offset / this.uploadFile['size']
+                        this.onUploadProcess.invoke(args)
+                    # end while
+                this.socket.send(packet.backup.uploadFinishRequest())
+            except Exception as e:
+                this.socket.send(packet.backup.uploadBreakRequest())
+        # end upload
+        
+        if not this.uploadFile:
+            this.lastOperationError = 'no avaiable upoad request.'
+            return False
+            
+        this.uploadInterrupted = False
+        this.uploadThread = threading.Thread(target = upload)
+        this.uploadThread.start()
+        result = this.waitForResponse()
+        
+        if not result:
+            this.uploadInterrupt()
+        return result
+    # end getUploadResult
+        
+    def uploadInterrupt(this):
+        this.socket.send(packet.backup.uploadBreakRequest())
+        if this.uploadThread:
+            this.uploadInterrupted = True
+            this.uploadThread.join()
+        this.uploadFile = { }
+        
     def downloadInterrupt(this):
         this.socket.send(packet.backup.downloadBreakRequest())
-        if this.downloadFile['file']:
+        if this.downloadFile and this.downloadFile['file']:
             this.downloadFile['file'].close()
             os.remove(this.downloadFile['temp'])
         this.downloadFile = { }
             
-    def downloadRequest(this, path):
+    def startDownload(this, path):
         temppath = ''
         while temppath == '' or os.path.exists(temppath):
             temppath = os.path.join(BAKSYS_TEMPDIR, 'download_%08X'%int(random()*0x100000000))
@@ -69,18 +138,21 @@ class BaksysRemoteBackup:
             'size'       : 0,                                         \
             'total-size' : 0                                          \
         }
-        if not os.path.exists(BAKSYS_TEMPDIR):
-            os.makedirs(BAKSYS_TEMPDIR)
-        this.downloadFile['file'] = open(this.downloadFile['temp'], 'wb')
         this.socket.send(packet.backup.downloadRequest(path))
-        return this.waitForResponse()
-    # end downloadRequest
+        result = this.waitForResponse()
+        if not result:
+            this.downloadInterrupt()
+        return result
+    # end startDownload
             
-    def downloadBackup(this):
+    def getDownloadResult(this):
         if not this.downloadFile:
             this.lastOperationError = 'no avaiable download request.'
             return False
-        return this.waitForResponse()
+        result = this.waitForResponse()
+        if not result:
+            this.downloadInterrupt()
+        return result
         
     def uploadBackup(this, backup, path):
         # TODO
